backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pymupdf
chromadb
openai
serpapi
sqlalchemy
python-dotenv
